refactor(auth): relocate auth modules into features/ directory for better structure
