--- src/app/api/auth/check-session/model.ts
@@ -1,5 +1,0 @@
-export type CheckSessionRequest = {
-  sessionId: string
-}
-
-export type CheckSessionResponse = { valid: boolean }
