--- src/app/api/auth/index.ts
@@ -1,5 +1,0 @@
-import { checkSession } from './check-session'
-
-export const Auth = {
-  checkSession,
-}
