--- src/app/api/index.ts
@@ -1,5 +1,0 @@
-import { Auth } from './auth'
-
-export const Api = {
-  Auth,
-}
