+++ src/app/login/page.tsx
@@ -1,107 +1,10 @@
-'use client'
-
-import { zodResolver } from '@hookform/resolvers/zod'
-import { CircleAlert } from 'lucide-react'
-import { useCallback, useState } from 'react'
-import { useForm } from 'react-hook-form'
-import zod from 'zod'
-
-import { login } from '@/app/actions/auth'
-import { Alert, AlertTitle, AlertDescription } from '@/components/ui/alert'
-import { Button } from '@/components/ui/button'
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
-import {
-  Form,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormControl,
-  FormMessage,
-} from '@/components/ui/form'
-import { Input } from '@/components/ui/input'
 import { Page } from '@/components/view/page'
-
-const schema = zod.object({
-  email: zod.email({ message: 'E-mail format is invalid' }),
-  password: zod.string().min(1, { message: 'Password must be not empty' }),
-})
-
-type FormValues = zod.infer<typeof schema>
+import { LoginCard } from '@/features/auth'
 
 export default function LoginPage() {
-  const [error, setError] = useState<string | null>(null)
-
-  const form = useForm<FormValues>({
-    resolver: zodResolver(schema),
-    defaultValues: {
-      email: '',
-      password: '',
-    },
-  })
-
-  const onSubmit = useCallback(async (values: FormValues) => {
-    setError(null)
-    const res = await login(values)
-    if (res?.error) {
-      setError(res.error)
-    }
-  }, [])
-
   return (
     <Page>
-      <Card className="w-full max-w-sm">
-        <CardHeader>
-          <CardTitle className="text-center text-2xl font-bold text-gray-800">Login</CardTitle>
-        </CardHeader>
-        <CardContent>
-          <Form {...form}>
-            <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6 p-8" noValidate>
-              <FormField
-                control={form.control}
-                name="email"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Email</FormLabel>
-                    <FormControl>
-                      <Input placeholder="Enter your email" type="email" {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name="password"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Password</FormLabel>
-                    <FormControl>
-                      <Input placeholder="Enter your password" type="password" {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <Button
-                type="submit"
-                className="w-full rounded-lg bg-blue-600 px-4 py-2 font-medium text-white transition hover:bg-blue-700"
-              >
-                Login
-              </Button>
-
-              {error && (
-                <Alert variant="destructive">
-                  <CircleAlert />
-                  <AlertTitle>Error!</AlertTitle>
-                  <AlertDescription>{error}</AlertDescription>
-                </Alert>
-              )}
-            </form>
-          </Form>
-        </CardContent>
-      </Card>
+      <LoginCard />
     </Page>
   )
 }
