+++ src/app/registerpage/page.tsx
@@ -1,12 +1,12 @@
 'use client'
 
 import { Page } from '@/components/view/page'
-import RegisterPageview from '@/components/view/register-page'
+import { RegisterCard } from '@/features/auth'
 
 export default function RegisterPage() {
   return (
     <Page>
-      <RegisterPageview />
+      <RegisterCard />
     </Page>
   )
 }
