+++ src/features/auth/actions.ts
@@ -1,35 +1,36 @@
 'use server'
 
-import { randomUUID } from 'crypto'
-
 import bcrypt from 'bcryptjs'
 import { cookies } from 'next/headers'
 import { redirect } from 'next/navigation'
 
-import { LoginPayload } from '@/definitions/payload/auth.model'
 import { prisma } from '@/lib/prisma'
+import { Err, SafePromiseResult } from '@/shared/safe-result'
 
-import { SESSION_COOKIE_NAME, SESSION_DURATION_DAYS } from './auth.const'
+import { SESSION_COOKIE_NAME, SESSION_DURATION_DAYS } from './const'
+import { LoginDto } from './dto'
 
-export async function login(data: LoginPayload) {
+export async function loginAction(
+  data: LoginDto.LoginRequest
+): SafePromiseResult<undefined, string> {
   const email = data.email
   const password = data.password
 
   if (!email || !password) {
-    return { error: 'Missing email or password' }
+    return Err('Missing email or password')
   }
 
   const user = await prisma.user.findUnique({ where: { email } })
   if (!user) {
-    return { error: 'Invalid email or password' }
+    return Err('Invalid email or password')
   }
 
   const isValid = await bcrypt.compare(password, user.passwordHash)
   if (!isValid) {
-    return { error: 'Invalid email or password' }
+    return Err('Invalid email or password')
   }
 
-  const sessionId = randomUUID()
+  const sessionId = crypto.randomUUID()
   const SevenDaysInMs = SESSION_DURATION_DAYS * 24 * 60 * 60 * 1000
   const expiresAt = new Date(Date.now() + SevenDaysInMs)
 
@@ -42,6 +43,7 @@ export async function login(data: LoginPayload) {
   })
 
   const cookieStore = await cookies()
+
   cookieStore.set(SESSION_COOKIE_NAME, sessionId, {
     httpOnly: true,
     secure: process.env.NODE_ENV === 'production',
