+++ src/features/auth/components/index.ts
@@ -1,0 +1,2 @@
+export * from './register-card'
+export * from './login-card'
