+++ src/features/auth/dto/index.ts
@@ -1,0 +1,3 @@
+import * as LoginDto from './login.dto'
+
+export { LoginDto }
