+++ src/features/auth/route/index.ts
@@ -1,0 +1,1 @@
+export * as CheckSessionRoute from '../route/check-session.route'
