+++ src/lib/middleware/auth.middleware.ts
@@ -1,13 +1,13 @@
 import { NextRequest, NextResponse } from 'next/server'
 
-import { checkSession } from '@/app/api/auth/check-session'
+import { checkSession } from '@/features/auth'
 
 async function protectDashboard(req: NextRequest) {
   const isDashboard = req.nextUrl.pathname.startsWith('/dashboard')
   if (!isDashboard) return null
 
-  const valid = await checkSession(req)
-  if (!valid) {
+  const res = await checkSession(req)
+  if (!res.ok || !res.value.valid) {
     return NextResponse.redirect(new URL('/login', req.url))
   }
 
@@ -18,8 +18,9 @@ async function preventMultiLogin(req: NextRequest) {
   const isLogin = req.nextUrl.pathname.startsWith('/login')
   if (!isLogin) return null
 
-  const valid = await checkSession(req)
-  if (valid) {
+  const res = await checkSession(req)
+
+  if (res.ok && res.value.valid) {
     return NextResponse.redirect(new URL('/dashboard', req.url))
   }
